Extracted library specific tests to separate executable.
