--- core/src/tests/generate_file.cpp
@@ -1,44 +1,0 @@
-/**********************************************************************************************************************
- * Copyright (c) 2021 Concurrent Technologies Corporation.                                                            *
- *                                                                                                                    *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance     *
- * with the License.  You may obtain a copy of the License at                                                         *
- *                                                                                                                    *
- *     http://www.apache.org/licenses/LICENSE-2.0                                                                     *
- *                                                                                                                    *
- * Unless required by applicable law or agreed to in writing, software is distributed under the License is            *
- * distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or                   *
- * implied.  See the License for the specific language governing permissions and limitations under the License.       *
- *                                                                                                                    *
- **********************************************************************************************************************/
-
-/**********************************************************************************************************************
- * Copyright (c) 2021 Concurrent Technologies Corporation.                                                            *
- *                                                                                                                    *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance     *
- * with the License.  You may obtain a copy of the License at                                                         *
- *                                                                                                                    *
- *     http://www.apache.org/licenses/LICENSE-2.0                                                                     *
- *                                                                                                                    *
- * Unless required by applicable law or agreed to in writing, software is distributed under the License is            *
- * distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or                   *
- * implied.  See the License for the specific language governing permissions and limitations under the License.       *
- *                                                                                                                    *
- **********************************************************************************************************************/
-
-#include "test_util.hpp"
-#include <cstdio>
-
-int main(int argc, char *argv[]) {
-    if (argc != 3) {
-        fprintf(stderr, "USAGE: %s num_bytes outfile\n", argv[0]);
-        return -1;
-    }
-    auto const fill = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
-                      ")!@#$%^&*(ΑαΒβΓγΔδΕεΖζΗηΘθΙιΚκΛλΜμΝνΞξΟοΠπΡρΣσςΤτΥυΦφΧχΨψΩω";
-    auto const fill_length = static_cast<int64_t>(strlen(fill));
-    auto const file_name = argv[2];
-    auto test_infile_ptr = fill_file(file_name, std::strtoll(argv[1], nullptr, 10), fill, fill_length);
-    fclose(test_infile_ptr);
-    return 0;
-}
