+++ core/src/tests/omegaEdit_tests.cpp
@@ -19,7 +19,9 @@
 #include "omega_edit/encode.h"
 #include "omega_edit/stl_string_adaptor.hpp"
 #include "omega_edit/utility.h"
-#include "test_util.hpp"
+
+#include <test_util.hpp>
+
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers_contains.hpp>
 #include <catch2/matchers/catch_matchers_string.hpp>
@@ -39,8 +41,6 @@ using Catch::Matchers::Contains;
 using Catch::Matchers::EndsWith;
 using Catch::Matchers::Equals;
 
-const auto DATA_DIR = std::filesystem::current_path() / "data";
-#define MAKE_PATH(path) (DATA_DIR / (path)).string().c_str()
 
 /**
  * Sleep for the given number of seconds.
@@ -63,526 +63,9 @@ TEST_CASE("Exported Constants", "[Constants]") {
     REQUIRE(sizeof(omega_byte_frequency_profile_t)/sizeof(int64_t) == omega_session_byte_frequency_profile_size());
 }
 
-TEST_CASE("Version check", "[VersionCheck]") {
-    const auto major = omega_version_major();
-    const auto minor = omega_version_minor();
-    const auto patch = omega_version_patch();
-    const auto version = (major << 24) + (minor << 16) + patch;
-    REQUIRE(0 < omega_version());
-    REQUIRE(version == omega_version());
-    REQUIRE(major == ((omega_version() >> 24) & 0xFF));
-    REQUIRE(minor == ((omega_version() >> 16) & 0xFF));
-    REQUIRE(patch == (omega_version() & 0xFF));
-    const auto libtype = omega_libtype();
-    REQUIRE(((strcmp("static", libtype) == 0) || (strcmp("shared", libtype) == 0)));
-}
-
-TEST_CASE("License check", "[LicenseCheck]") {
-    const auto license = omega_license_get();
-    REQUIRE(license);
-    REQUIRE(strlen(license) == 576);
-    REQUIRE(strstr(license, "Concurrent Technologies Corporation"));
-}
-
 TEST_CASE("Bit Manipulation", "[BitManip]") {
     REQUIRE(1 << 31 == -2147483648);
     REQUIRE(~(1 << 31) == 2147483647);
-}
-
-TEST_CASE("Buffer Shift", "[BufferShift]") {
-    auto const fill = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-    auto const buff_len = (int64_t) strlen(fill);
-    auto *buffer = (omega_byte_t *) omega_util_strndup(fill, buff_len);
-
-    // Negative tests.  Shifting 8 or more bits in either direction should be an error.
-    REQUIRE(-1 == omega_util_left_shift_buffer(buffer, buff_len, 8, 0));
-    REQUIRE(-1 == omega_util_right_shift_buffer(buffer, buff_len, 8, 0));
-    REQUIRE(-1 == omega_util_left_shift_buffer(buffer, buff_len, 0, 1));
-    REQUIRE(-1 == omega_util_right_shift_buffer(buffer, buff_len, 0, 1));
-    REQUIRE(-2 == omega_util_left_shift_buffer(buffer, buff_len, 4, 2));
-    REQUIRE(-2 == omega_util_right_shift_buffer(buffer, buff_len, 4, 2));
-
-    // Shift the buffer 3 bits to the right
-    REQUIRE(0 == omega_util_right_shift_buffer(buffer, buff_len, 3, 0));
-    // Shift the buffer 5 bits to the right
-    REQUIRE(0 == omega_util_right_shift_buffer(buffer, buff_len, 5, 0));
-    // Should equal \x000123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxy
-    // We shifted a total of 8 bits (one byte) to the right, so compare the first byte of the buffer with null and the buffer plus one against the fill minus the first character
-    REQUIRE(buffer[0] == '\x00');
-    REQUIRE_THAT(std::string((const char *) buffer + 1), Equals(std::string(fill, buff_len - 1)));
-
-    // Reset the buffer
-    strcpy((char *) buffer, fill);
-
-    // Shift the buffer 3 bits to the right, filling the vacancies with 1s
-    REQUIRE(0 == omega_util_right_shift_buffer(buffer, buff_len, 3, 1));
-    // Shift the buffer 5 bits to the right, filling the vacancies with 1s
-    REQUIRE(0 == omega_util_right_shift_buffer(buffer, buff_len, 5, 1));
-    // Should equal \xFF0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxy
-    // We shifted a total of 8 bits (one byte) to the right, so compare the buffer against the fill minus the last character, with the first byte filled with 1s
-    REQUIRE_THAT(std::string((const char *) buffer), Equals(std::string("\xFF") + std::string(fill, buff_len - 1)));
-
-    // Reset the buffer
-    strcpy((char *) buffer, fill);
-
-    // Shift the buffer 2 bits to the left
-    REQUIRE(0 == omega_util_left_shift_buffer(buffer, buff_len, 2, 0));
-    // Shift the buffer 6 bits to the left
-    REQUIRE(0 == omega_util_left_shift_buffer(buffer, buff_len, 6, 0));
-    // Should equal 123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz\x00
-    // We shifted a total of 8 bits (one byte) to the left, so compare the buffer against the fill plus one null byte
-    REQUIRE(buffer[buff_len] == '\x00');
-    REQUIRE_THAT(std::string((const char *) buffer, buff_len - 1), Equals(std::string(fill + 1, buff_len - 1)));
-
-    // Left shift tests with fill bit 1
-    // Reset the buffer
-    strcpy((char *) buffer, fill);
-
-    // Shift the buffer 2 bits to the left, filling the vacancies with 1s
-    REQUIRE(0 == omega_util_left_shift_buffer(buffer, buff_len, 2, 1));
-    // Shift the buffer 6 bits to the left, filling the vacancies with 1s
-    REQUIRE(0 == omega_util_left_shift_buffer(buffer, buff_len, 6, 1));
-    // Should equal 123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz\xFF
-    // We shifted a total of 8 bits (one byte) to the left, so compare the buffer against the fill minus the first character, with the last byte filled with 1s
-    REQUIRE_THAT(std::string((const char *) buffer), Equals(std::string(fill + 1) + "\xFF"));
-
-    strcpy((char *) buffer, "ABCD");
-    REQUIRE(0 == omega_util_right_shift_buffer(buffer, 4, 2, 0));
-    REQUIRE_THAT(std::string((const char *) buffer), Equals("\x10\x50\x90\xD1"));
-
-    strcpy((char *) buffer, "WXYZ");
-    REQUIRE(0 == omega_util_right_shift_buffer(buffer, 4, 4, 1));
-    REQUIRE_THAT(std::string((const char *) buffer), Equals("\xF5\x75\x85\x95"));
-
-    strcpy((char *) buffer, "1234");
-    REQUIRE(0 == omega_util_left_shift_buffer(buffer, 4, 3, 0));
-    REQUIRE_THAT(std::string((const char *) buffer), Equals("\x89\x91\x99\xA0"));
-
-    strcpy((char *) buffer, "abcd");
-    REQUIRE(0 == omega_util_left_shift_buffer(buffer, 4, 7, 1));
-    REQUIRE_THAT(std::string((const char *) buffer), Equals("\xB1\x31\xB2\x7F"));
-
-    free(buffer);
-}
-
-TEST_CASE("File Compare", "[UtilTests]") {
-    SECTION("Identity") {
-        // Same file ought to yield identical contents
-        REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.dat"), MAKE_PATH("test1.dat")));
-    }
-    SECTION("Difference") {
-        // Different files with different contents
-        REQUIRE(0 != omega_util_compare_files(MAKE_PATH("test1.dat"), MAKE_PATH("test2.dat")));
-    }
-}
-
-TEST_CASE("File Copy", "[UtilTests]") {
-    struct stat src_stat{};
-    struct stat dst_stat{};
-    omega_util_remove_file(MAKE_PATH("test1.copy.dat"));
-#ifdef OMEGA_BUILD_WINDOWS
-    // sleep for 1 second to ensure the file modification time is different,
-    // needed for Windows because the file system only has 100 nanosecond resolution
-    omega_util_sleep_(1);
-#endif
-    REQUIRE(0 == omega_util_file_copy(MAKE_PATH("test1.dat"), MAKE_PATH("test1.copy.dat"), 0));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.dat"), MAKE_PATH("test1.copy.dat")));
-
-    REQUIRE(0 == omega_util_compare_modification_times(MAKE_PATH("test1.dat"), MAKE_PATH("test1.dat")));
-    REQUIRE(1 == omega_util_compare_modification_times(MAKE_PATH("test1.copy.dat"), MAKE_PATH("test1.dat")));
-    REQUIRE(-1 == omega_util_compare_modification_times(MAKE_PATH("test1.dat"), MAKE_PATH("test1.copy.dat")));
-    REQUIRE(-2 == omega_util_compare_modification_times(MAKE_PATH("test1.dat"), "-invalid-"));
-
-    REQUIRE(0 == stat(MAKE_PATH("test1.dat"), &src_stat));
-    REQUIRE(0 == stat(MAKE_PATH("test1.copy.dat"), &dst_stat));
-
-    // The mode includes the file type
-    const int dst_mode = 0100600;// S_IFREG | S_IRUSR  regular file with owner read-only
-    REQUIRE(dst_mode != src_stat.st_mode);
-    REQUIRE(src_stat.st_mode == dst_stat.st_mode);
-
-    REQUIRE(0 == omega_util_remove_file(MAKE_PATH("test1.copy.dat")));
-    REQUIRE(0 == omega_util_file_copy(MAKE_PATH("test1.dat"), MAKE_PATH("test1.copy.dat"), dst_mode));
-
-    REQUIRE(0 == stat((DATA_DIR / "test1.copy.dat").string().c_str(), &dst_stat));
-
-#ifndef OMEGA_BUILD_WINDOWS
-    // On Windows, the mode is not preserved as expected on non-Windows platforms
-    REQUIRE(dst_mode == dst_stat.st_mode);
-#endif
-
-    REQUIRE(omega_util_directory_exists(DATA_DIR.string().c_str()));
-    REQUIRE(!omega_util_file_exists(DATA_DIR.string().c_str()));
-    REQUIRE(!omega_util_directory_exists((DATA_DIR / "test1.copy.dat").string().c_str()));
-    REQUIRE(omega_util_file_exists((DATA_DIR / "test1.copy.dat").string().c_str()));
-    REQUIRE(1 == omega_util_remove_all((DATA_DIR / "test1.copy.dat").string().c_str()));
-    REQUIRE(0 == omega_util_remove_all((DATA_DIR / "test1.copy.dat").string().c_str()));
-    REQUIRE(!omega_util_file_exists((DATA_DIR / "test1.copy.dat").string().c_str()));
-}
-
-TEST_CASE("End Of Line", "[EOLTests]") {
-    omega_byte_t buffer[1024];
-    FILE *in_fp = FOPEN(MAKE_PATH("test1.dat"), "rb");
-    REQUIRE(in_fp);
-    auto file_size = omega_util_file_size(MAKE_PATH("test1.dat"));
-    REQUIRE(63 == file_size);
-    REQUIRE(file_size < sizeof(buffer));
-    REQUIRE(file_size == fread(buffer, sizeof(omega_byte_t), file_size, in_fp));
-    REQUIRE(0 == FCLOSE(in_fp));
-    FILE *out_fp = FOPEN(MAKE_PATH("test1.actual.eol.1.dat"), "wb");
-    REQUIRE(out_fp);
-    REQUIRE(file_size == fwrite(buffer, sizeof(omega_byte_t), file_size, out_fp));
-    REQUIRE(file_size == FTELL(out_fp));
-    REQUIRE(0 == FCLOSE(out_fp));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.dat"), MAKE_PATH("test1.actual.eol.1.dat")));
-}
-
-TEST_CASE("File Exists", "[UtilTests]") {
-    REQUIRE(fs::exists(MAKE_PATH("test1.dat")));
-    omega_util_remove_file(MAKE_PATH("IDonTExist.DaT"));
-    REQUIRE(!fs::exists(MAKE_PATH("IDonTExist.DaT")));
-    REQUIRE(0 != omega_util_file_exists(MAKE_PATH("test1.dat")));
-    REQUIRE(0 == omega_util_file_exists(MAKE_PATH("IDonTExist.DaT")));
-}
-
-TEST_CASE("File Touch", "[UtilTests]") {
-    const char dir_sep = omega_util_directory_separator();
-    const auto exists = std::string(MAKE_PATH("test1.dat"));
-    const auto dont_exist = std::string(MAKE_PATH("IDonTExistYeT.DaT"));
-    omega_util_remove_file(MAKE_PATH("IDonTExistYeT.DaT"));
-    REQUIRE(omega_util_file_exists(exists.c_str()));
-    REQUIRE(!omega_util_file_exists(dont_exist.c_str()));
-    auto expected = std::string(MAKE_PATH("test1-1.dat"));
-    REQUIRE_THAT(omega_util_available_filename(exists.c_str(), nullptr), Equals(expected));
-    expected = dont_exist;
-    REQUIRE_THAT(omega_util_available_filename(dont_exist.c_str(), nullptr), Equals(expected));
-    omega_util_touch(dont_exist.c_str(),
-                     0);// logs an error as expected because create is false and the file does not exist
-    REQUIRE(!omega_util_file_exists(dont_exist.c_str()));
-#ifdef OMEGA_BUILD_WINDOWS
-    // sleep for 1 second to ensure the file modification time is different,
-    // needed for Windows because the file system only has 100 nanosecond resolution
-    omega_util_sleep_(1);
-#endif
-    omega_util_touch(dont_exist.c_str(), 1);
-    REQUIRE(omega_util_file_exists(dont_exist.c_str()));
-    REQUIRE(-1 == omega_util_compare_modification_times(exists.c_str(), dont_exist.c_str()));
-    REQUIRE(1 == omega_util_compare_modification_times(dont_exist.c_str(), exists.c_str()));
-    expected = std::string(MAKE_PATH("IDonTExistYeT-1.DaT"));
-    REQUIRE_THAT(omega_util_available_filename(dont_exist.c_str(), nullptr), Equals(expected));
-    REQUIRE(0 == omega_util_remove_file(dont_exist.c_str()));
-    REQUIRE(!omega_util_file_exists(dont_exist.c_str()));
-    expected = dont_exist;
-    REQUIRE_THAT(omega_util_available_filename(dont_exist.c_str(), nullptr), Equals(expected));
-}
-
-TEST_CASE("Current Directory", "[UtilTests]") {
-    const auto current_directory = omega_util_get_current_dir(nullptr);
-    REQUIRE(current_directory);
-    REQUIRE(omega_util_directory_exists(current_directory));
-}
-
-TEST_CASE("Directory Name", "[UtilTests]") {
-    // Unix-style paths
-    auto test_1 = "/this/is/a/directory/filename.extension";
-    char buffer[FILENAME_MAX];
-    auto result = omega_util_dirname(test_1, nullptr);
-    REQUIRE(result);
-    REQUIRE_THAT(result, Equals("/this/is/a/directory"));
-    // DOS/Windows-style paths
-    auto test_2 = R"(C:\this\is\a\directory\filename.extension)";
-    result = omega_util_dirname(test_2, buffer);
-    REQUIRE(result);
-#ifdef OMEGA_BUILD_WINDOWS
-    REQUIRE_THAT(buffer, Equals(R"(C:\this\is\a\directory)"));
-#else
-    REQUIRE_THAT(buffer, Equals(""));
-#endif
-    // Missing directory test
-    auto test_3 = "filename.extension";
-    result = omega_util_dirname(test_3, buffer);
-    REQUIRE(result);
-    REQUIRE_THAT(buffer, Equals(""));
-    // relative path
-    auto test_4 = "relative/filename.extension";
-    result = omega_util_dirname(test_4, buffer);
-    REQUIRE(result);
-    REQUIRE_THAT(buffer, Equals("relative"));
-}
-
-TEST_CASE("Base File Name", "[UtilTests]") {
-    // Unix-style paths
-    auto test_1 = "/this/is/a/directory/filename.extension";
-    char buffer[FILENAME_MAX];
-    auto result = omega_util_basename(test_1, nullptr, 0);
-    REQUIRE(result);
-    REQUIRE_THAT(result, Equals("filename.extension"));
-    // DOS/Windows-style paths
-    auto test_2 = R"(C:\this\is\a\directory\filename.extension)";
-    result = omega_util_basename(test_2, buffer, 0);
-    REQUIRE(result);
-#ifdef OMEGA_BUILD_WINDOWS
-    REQUIRE_THAT(buffer, Equals("filename.extension"));
-#else
-    REQUIRE_THAT(buffer, Equals("C:\\this\\is\\a\\directory\\filename.extension"));
-#endif
-    auto test_3 = "filename.extension";
-    result = omega_util_basename(test_3, buffer, 0);
-    REQUIRE(result);
-    REQUIRE_THAT(buffer, Equals("filename.extension"));
-    result = omega_util_basename(test_3, buffer, 1);
-    REQUIRE(result);
-    REQUIRE_THAT(buffer, Equals("filename"));
-    auto test_4 = "/this/is/a/directory/";
-    result = omega_util_basename(test_4, buffer, 0);
-    REQUIRE(result);
-    REQUIRE_THAT(buffer, Equals(""));
-}
-
-TEST_CASE("File Extension", "[UtilTests]") {
-    // Unix-style paths
-    const auto test_1 = "/this/is/a/directory/filename.extension";
-    char buffer[FILENAME_MAX];
-    auto result = omega_util_file_extension(test_1, nullptr);
-    REQUIRE(result);
-    REQUIRE_THAT(result, Equals(".extension"));
-    // DOS/Windows-style paths
-    const auto test_2 = R"(C:\this\is\a\directory\filename.extension)";
-    result = omega_util_file_extension(test_2, buffer);
-    REQUIRE(result);
-    REQUIRE_THAT(buffer, Equals(".extension"));
-    const auto test_3 = "filename_no_extension";
-    result = omega_util_file_extension(test_3, buffer);
-    REQUIRE_THAT(result, Equals(""));
-    const auto test_4 = "filename_empty_extension.";
-    result = omega_util_file_extension(test_4, buffer);
-    REQUIRE(result);
-    REQUIRE_THAT(result, Equals("."));
-    const auto test_5 = "/..";
-    result = omega_util_file_extension(test_5, buffer);
-    REQUIRE(result);
-    REQUIRE_THAT(result, Equals(""));
-    const auto test_6 = "/this.is.a.directory/filename_no_extension";
-    result = omega_util_file_extension(test_6, buffer);
-    REQUIRE_THAT(result, Equals(""));
-}
-
-static inline omega_byte_t to_lower(omega_byte_t byte, void *) { return tolower(byte); }
-
-static inline omega_byte_t to_upper(omega_byte_t byte, void *) { return toupper(byte); }
-
-TEST_CASE("Transformer", "[TransformerTest]") {
-    omega_byte_t bytes[32];
-    strcpy(reinterpret_cast<char *>(bytes), "Hello World!");
-    const auto bytes_length = static_cast<int64_t>(strlen(reinterpret_cast<const char *>(bytes)));
-    omega_util_apply_byte_transform(bytes, bytes_length, to_upper, nullptr);
-    REQUIRE_THAT(string(reinterpret_cast<const char *>(bytes)), Equals("HELLO WORLD!"));
-    omega_util_apply_byte_transform(bytes, bytes_length, to_lower, nullptr);
-    REQUIRE_THAT(string(reinterpret_cast<const char *>(bytes)), Equals("hello world!"));
-    omega_util_apply_byte_transform(bytes, 1, to_upper, nullptr);
-    REQUIRE_THAT(string(reinterpret_cast<const char *>(bytes)), Equals("Hello world!"));
-}
-
-TEST_CASE("File Transformer", "[TransformerTest]") {
-    REQUIRE(0 == omega_util_apply_byte_transform_to_file(
-        MAKE_PATH("test1.dat"), MAKE_PATH("test1.actual.transformed.1.dat"), to_upper, nullptr, 0, 0));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.expected.transformed.1.dat"),
-        MAKE_PATH("test1.actual.transformed.1.dat")));
-    REQUIRE(0 == omega_util_apply_byte_transform_to_file(MAKE_PATH("test1.dat"),
-        MAKE_PATH("test1.actual.transformed.2.dat"), to_lower, nullptr,
-        37, 10));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.expected.transformed.2.dat"),
-        MAKE_PATH("test1.actual.transformed.2.dat")));
-    REQUIRE(0 != omega_util_apply_byte_transform_to_file(MAKE_PATH("test1.dat"),
-        MAKE_PATH("test1.actual.transformed.3.dat"), to_lower, nullptr,
-        37, 100));
-    REQUIRE(0 == omega_util_file_exists(MAKE_PATH("test1.actual.transformed.3.dat")));
-}
-
-TEST_CASE("Encoding", "[EncodingTest]") {
-    auto in_string = string("Hello World!");
-    auto in = reinterpret_cast<const omega_byte_t *>(in_string.data());
-    char encoded_buffer[1024];
-    omega_byte_t decoded_buffer[1024];
-    omega_encode_bin2hex(in, encoded_buffer, in_string.size());
-    REQUIRE_THAT(encoded_buffer, Equals("48656c6c6f20576f726c6421"));
-    omega_encode_hex2bin(encoded_buffer, decoded_buffer, strlen(encoded_buffer));
-    REQUIRE_THAT(reinterpret_cast<const char *>(decoded_buffer), Equals(in_string));
-    omega_encode_hex2bin("48656C6C6F20576F726C6421", decoded_buffer, strlen(encoded_buffer));
-    REQUIRE_THAT(reinterpret_cast<const char *>(decoded_buffer), Equals(in_string));
-}
-
-using file_info_t = struct file_info_struct {
-    size_t num_changes{};
-};
-
-static inline void session_change_cbk(const omega_session_t *session_ptr, omega_session_event_t session_event,
-                                      const void *session_event_ptr) {
-    // Not all session changes are the result of a standard change like delete / insert / overwrite
-    switch (session_event) {
-        case SESSION_EVT_EDIT:
-        case SESSION_EVT_UNDO: {
-            auto file_info_ptr = reinterpret_cast<file_info_t *>(omega_session_get_user_data_ptr(session_ptr));
-            const auto change_ptr = reinterpret_cast<const omega_change_t *>(session_event_ptr);
-            const auto bytes = omega_change_get_bytes(change_ptr);
-            const auto bytes_length = omega_change_get_length(change_ptr);
-            if (0 < omega_change_get_serial(change_ptr)) { ++file_info_ptr->num_changes; } else {
-                --file_info_ptr->num_changes;/* this is an UNDO */
-            }
-            auto file_path = omega_session_get_file_path(session_ptr);
-            file_path = (file_path) ? file_path : "NO FILENAME";
-            clog << dec << R"({ "filename" : ")" << file_path << R"(", "num_changes" : )"
-                    << omega_session_get_num_changes(session_ptr) << R"(, "computed_file_size": )"
-                    << omega_session_get_computed_file_size(session_ptr) << R"(, "change_transaction_bit": )"
-                    << omega_change_get_transaction_bit(change_ptr) << R"(, "change_serial": )"
-                    << omega_change_get_serial(change_ptr) << R"(, "change_kind": ")"
-                    << omega_change_get_kind_as_char(change_ptr) << R"(", "offset": )"
-                    << omega_change_get_offset(change_ptr) << R"(, "length": )" << omega_change_get_length(change_ptr);
-            if (bytes) { clog << R"(, "bytes": ")" << string((const char *) bytes, bytes_length) << R"(")"; }
-            clog << "}" << endl;
-        }
-        default:
-            break;
-    }
-}
-
-TEST_CASE("Empty File Tests", "[EmptyFileTests]") {
-    file_info_t file_info;
-    file_info.num_changes = 0;
-    const std::string in_filename_str(MAKE_PATH("empty_file.dat"));
-    const auto in_filename = in_filename_str.c_str();
-    auto file_size = omega_util_file_size(in_filename);
-    REQUIRE(0 == file_size);
-    REQUIRE(!getenv("OMEGA_EDIT_CHECKPOINT_DIRECTORY"));// make sure this environment variable is not set
-    const auto session_ptr = omega_edit_create_session(in_filename, session_change_cbk, &file_info,
-                                                       SESSION_EVT_EDIT | SESSION_EVT_UNDO, nullptr);
-    REQUIRE(session_ptr);
-    REQUIRE_THAT(omega_session_get_file_path(session_ptr), Equals(in_filename));
-
-    REQUIRE(omega_util_paths_equivalent(omega_session_get_checkpoint_directory(session_ptr),
-                omega_util_dirname(in_filename, nullptr))
-
-            );
-    REQUIRE(omega_session_get_computed_file_size(session_ptr) == file_size);
-    REQUIRE(0 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(0 == omega_session_get_num_change_transactions(session_ptr));
-    REQUIRE(0 == omega_edit_undo_last_change(session_ptr));
-    auto change_serial =
-            omega_edit_insert_bytes(session_ptr, 0, reinterpret_cast<const omega_byte_t *>("1234567890"), 0);
-    REQUIRE(0 < change_serial);
-    REQUIRE(1 == omega_session_get_num_change_transactions(session_ptr));
-    file_size += 10;
-    REQUIRE(file_size == omega_session_get_computed_file_size(session_ptr));
-    REQUIRE(1 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(0 == omega_session_get_num_undone_change_transactions(session_ptr));
-    REQUIRE((change_serial * -1) == omega_edit_undo_last_change(session_ptr));
-    REQUIRE(0 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(1 == omega_session_get_num_undone_change_transactions(session_ptr));
-    REQUIRE(0 == omega_session_get_computed_file_size(session_ptr));
-    change_serial = omega_edit_overwrite_string(session_ptr, 0, "abcdefghhijklmnopqrstuvwxyz");
-    REQUIRE(0 < change_serial);
-    REQUIRE(0 == omega_session_get_num_undone_changes(session_ptr));
-    REQUIRE(0 == omega_session_get_num_undone_change_transactions(session_ptr));
-    REQUIRE(1 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(1 == omega_session_get_num_change_transactions(session_ptr));
-    REQUIRE(27 == omega_session_get_computed_file_size(session_ptr));
-    omega_edit_destroy_session(session_ptr);
-}
-
-typedef struct mask_info_struct {
-    omega_byte_t mask;
-    omega_mask_kind_t mask_kind;
-} mask_info_t;
-
-static inline omega_byte_t byte_mask_transform(omega_byte_t byte, void *user_data_ptr) {
-    const auto mask_info_ptr = reinterpret_cast<mask_info_t *>(user_data_ptr);
-    return omega_util_mask_byte(byte, mask_info_ptr->mask, mask_info_ptr->mask_kind);
-}
-
-TEST_CASE("Checkpoint Tests", "[CheckpointTests]") {
-    file_info_t file_info;
-    file_info.num_changes = 0;
-    const auto in_filename_str = std::string(MAKE_PATH("test1.dat"));
-    const auto in_filename = in_filename_str.c_str();
-    const auto session_ptr =
-            omega_edit_create_session(in_filename, session_change_cbk, &file_info, ALL_EVENTS, nullptr);
-    REQUIRE(session_ptr);
-    auto file_size = omega_session_get_computed_file_size(session_ptr);
-    REQUIRE(file_size > 0);
-    REQUIRE(0 !=
-            omega_edit_insert_string(session_ptr, 0, "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"));
-    REQUIRE(1 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(0 == omega_session_get_num_checkpoints(session_ptr));
-    REQUIRE(-1 == omega_edit_destroy_last_checkpoint(session_ptr));
-    REQUIRE(0 == omega_edit_apply_transform(session_ptr, to_lower, nullptr, 0, 0));
-    REQUIRE(1 == omega_session_get_num_checkpoints(session_ptr));
-    REQUIRE(1 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(1 == omega_session_get_num_change_transactions(session_ptr));
-    REQUIRE(2 == omega_edit_overwrite_string(session_ptr, 37, "BCDEFGHIJKLMNOPQRSTUVWXY"));
-    REQUIRE(2 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(2 == omega_session_get_num_change_transactions(session_ptr));
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("test1.actual.checkpoint.1.dat"),
-        omega_io_flags_t::IO_FLG_OVERWRITE, nullptr));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.expected.checkpoint.1.dat"),
-        MAKE_PATH("test1.actual.checkpoint.1.dat")));
-    mask_info_t mask_info;
-    mask_info.mask_kind = MASK_XOR;
-    mask_info.mask = 0xFF;
-    REQUIRE(0 == omega_edit_apply_transform(session_ptr, byte_mask_transform, &mask_info, 10, 26));
-    REQUIRE(2 == omega_session_get_num_checkpoints(session_ptr));
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("test1.actual.checkpoint.2.dat"),
-        omega_io_flags_t::IO_FLG_OVERWRITE, nullptr));
-    REQUIRE(0 == omega_edit_apply_transform(session_ptr, byte_mask_transform, &mask_info, 10, 26));
-    REQUIRE(3 == omega_session_get_num_checkpoints(session_ptr));
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("test1.actual.checkpoint.3.dat"),
-        omega_io_flags_t::IO_FLG_OVERWRITE, nullptr));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.expected.checkpoint.1.dat"),
-        MAKE_PATH("test1.actual.checkpoint.3.dat")));
-    mask_info.mask_kind = MASK_AND;
-    REQUIRE(0 == omega_edit_apply_transform(session_ptr, byte_mask_transform, &mask_info, 10, 0));
-    REQUIRE(4 == omega_session_get_num_checkpoints(session_ptr));
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("test1.actual.checkpoint.4.dat"),
-        omega_io_flags_t::IO_FLG_OVERWRITE, nullptr));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.expected.checkpoint.1.dat"),
-        MAKE_PATH("test1.actual.checkpoint.4.dat")));
-    mask_info.mask_kind = MASK_OR;
-    mask_info.mask = 0x00;
-    REQUIRE(0 == omega_edit_apply_transform(session_ptr, byte_mask_transform, &mask_info, 10, 0));
-    REQUIRE(5 == omega_session_get_num_checkpoints(session_ptr));
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("test1.actual.checkpoint.5.dat"),
-        omega_io_flags_t::IO_FLG_OVERWRITE, nullptr));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.expected.checkpoint.1.dat"),
-        MAKE_PATH("test1.actual.checkpoint.5.dat")));
-    mask_info.mask_kind = MASK_AND;
-    REQUIRE(0 == omega_edit_apply_transform(session_ptr, byte_mask_transform, &mask_info, 10, 0));
-    REQUIRE(6 == omega_session_get_num_checkpoints(session_ptr));
-    REQUIRE(3 == omega_edit_overwrite_string(session_ptr, 0,
-        "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"));
-    REQUIRE(3 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(3 == omega_session_get_num_change_transactions(session_ptr));
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("test1.actual.checkpoint.6.dat"),
-        omega_io_flags_t::IO_FLG_OVERWRITE, nullptr));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.expected.checkpoint.6.dat"),
-        MAKE_PATH("test1.actual.checkpoint.6.dat")));
-    auto change_ptr = omega_session_get_last_change(session_ptr);
-    REQUIRE(change_ptr);
-    REQUIRE(3 == omega_change_get_serial(change_ptr));
-    REQUIRE(4 == omega_edit_insert_string(session_ptr, 0, "12345"));
-    REQUIRE(5 == omega_edit_delete(session_ptr, 0, 5));
-    REQUIRE(5 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(5 == omega_session_get_num_change_transactions(session_ptr));
-    change_ptr = omega_session_get_last_change(session_ptr);
-    REQUIRE(5 == omega_change_get_serial(change_ptr));
-    REQUIRE(0 == omega_edit_destroy_last_checkpoint(session_ptr));
-    REQUIRE(5 == omega_session_get_num_checkpoints(session_ptr));
-    REQUIRE(2 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(2 == omega_session_get_num_change_transactions(session_ptr));
-    REQUIRE(nullptr == omega_session_get_last_change(session_ptr));
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("test1.actual.checkpoint.7.dat"),
-        omega_io_flags_t::IO_FLG_OVERWRITE, nullptr));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("test1.expected.checkpoint.1.dat"),
-        MAKE_PATH("test1.actual.checkpoint.7.dat")));
-    omega_edit_destroy_session(session_ptr);
 }
 
 TEST_CASE("Model Tests", "[ModelTests]") {
@@ -1522,293 +1005,3 @@ TEST_CASE("Viewports", "[ViewportTests]") {
     omega_edit_destroy_session(session_ptr);
 }
 
-void session_save_test_session_cbk(const omega_session_t *session_ptr, omega_session_event_t session_event,
-                                   const void *) {
-    auto count_ptr = reinterpret_cast<int *>(omega_session_get_user_data_ptr(session_ptr));
-    std::clog << "Session Event: " << session_event << std::endl;
-    ++*count_ptr;
-}
-
-void session_save_test_viewport_cbk(const omega_viewport_t *viewport_ptr, omega_viewport_event_t viewport_event,
-                                    const void *) {
-    auto count_ptr = reinterpret_cast<int *>(omega_viewport_get_user_data_ptr(viewport_ptr));
-    std::clog << "Viewport Event: " << viewport_event << std::endl;
-    ++*count_ptr;
-}
-
-TEST_CASE("Session Save", "[SessionSaveTests]") {
-    char saved_filename[FILENAME_MAX];
-    int session_events_count = 0;
-    int viewport_events_count = 0;
-    auto session_ptr = omega_edit_create_session(nullptr, session_save_test_session_cbk, &session_events_count,
-                                                 ALL_EVENTS, nullptr);
-    REQUIRE(1 == session_events_count);// SESSION_EVT_CREATE
-    auto viewport_ptr = omega_edit_create_viewport(session_ptr, 0, 100, 0, session_save_test_viewport_cbk,
-                                                   &viewport_events_count, ALL_EVENTS);
-
-    REQUIRE(0 == omega_viewport_get_following_byte_count(viewport_ptr));
-    REQUIRE(2 == session_events_count);// SESSION_EVT_CREATE_VIEWPORT
-    REQUIRE(1 == viewport_events_count);// VIEWPORT_EVT_CREATE
-    omega_edit_insert_string(session_ptr, 0, "0123456789");
-    REQUIRE(3 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(2 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    omega_util_remove_file(MAKE_PATH("session_save.1.dat"));
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("session_save.1.dat"), omega_io_flags_t::IO_FLG_OVERWRITE,
-        saved_filename));
-    REQUIRE(omega_util_paths_equivalent(MAKE_PATH("session_save.1.dat"), saved_filename));
-    REQUIRE(0 == omega_edit_save_segment(session_ptr, MAKE_PATH("session_save_seg.1.dat"),
-        omega_io_flags_t::IO_FLG_OVERWRITE, saved_filename, 1, 0));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("session_save_seg.expected.1.dat"),
-        MAKE_PATH("session_save_seg.1.dat")));
-    omega_util_remove_file(MAKE_PATH("session_save_seg.1.dat"));
-    REQUIRE(0 == omega_edit_save_segment(session_ptr, MAKE_PATH("session_save_seg.2.dat"),
-        omega_io_flags_t::IO_FLG_OVERWRITE, saved_filename, 0, 4));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("session_save_seg.expected.2.dat"),
-        MAKE_PATH("session_save_seg.2.dat")));
-    omega_util_remove_file(MAKE_PATH("session_save_seg.2.dat"));
-    REQUIRE(0 == omega_edit_save_segment(session_ptr, MAKE_PATH("session_save_seg.3.dat"),
-        omega_io_flags_t::IO_FLG_OVERWRITE, saved_filename, 2, 6));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("session_save_seg.expected.3.dat"),
-        MAKE_PATH("session_save_seg.3.dat")));
-    omega_util_remove_file(MAKE_PATH("session_save_seg.3.dat"));
-    REQUIRE(7 == session_events_count);// SESSION_EVT_SAVE
-    REQUIRE(2 == viewport_events_count);// no additional viewport events
-    omega_edit_destroy_session(session_ptr);
-    session_events_count = 0;
-    viewport_events_count = 0;
-    session_ptr = omega_edit_create_session(MAKE_PATH("session_save.1.dat"), session_save_test_session_cbk,
-                                            &session_events_count, ALL_EVENTS, nullptr);
-    REQUIRE(1 == session_events_count);
-    viewport_ptr = omega_edit_create_viewport(session_ptr, 0, 100, 0, session_save_test_viewport_cbk,
-                                              &viewport_events_count, ALL_EVENTS);
-    REQUIRE(2 == session_events_count);
-    REQUIRE(1 == viewport_events_count);
-    REQUIRE(0 == omega_viewport_get_following_byte_count(viewport_ptr));
-    omega_edit_insert_string(session_ptr, omega_session_get_computed_file_size(session_ptr),
-                             "abcdefghijklmnopqrstuvwxyz");
-    REQUIRE(1 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(3 == session_events_count);
-    REQUIRE(2 == viewport_events_count);
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("session_save.1.dat"), omega_io_flags_t::IO_FLG_OVERWRITE,
-        saved_filename));
-    REQUIRE(omega_util_paths_equivalent(MAKE_PATH("session_save.1.dat"), saved_filename));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("session_save.expected.1.dat"), MAKE_PATH("session_save.1.dat")));
-    REQUIRE(1 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(4 == session_events_count);// SESSION_EVT_SAVE
-    REQUIRE(2 == viewport_events_count);// no additional viewport events
-    omega_edit_insert_string(session_ptr, omega_session_get_computed_file_size(session_ptr),
-                             "ABCDEFGHIJKLMNOPQRSTUVWXYZ");
-    REQUIRE(2 == omega_session_get_num_changes(session_ptr));
-    omega_util_remove_file(MAKE_PATH("session_save.1-1.dat"));
-    REQUIRE(5 == session_events_count);// SESSION_EVT_SAVE
-    REQUIRE(3 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("session_save.1.dat"), omega_io_flags_t::IO_FLG_NONE,
-        saved_filename));
-    REQUIRE(6 == session_events_count);// SESSION_EVT_SAVE
-    REQUIRE(omega_util_paths_equivalent(MAKE_PATH("session_save.1-1.dat"), saved_filename));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("session_save.expected.2.dat"), MAKE_PATH("session_save.1-1.dat")));
-    omega_util_remove_file(MAKE_PATH("session_save.1-2.dat"));
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("session_save.1.dat"), omega_io_flags_t::IO_FLG_NONE,
-        saved_filename));
-    REQUIRE(7 == session_events_count);// SESSION_EVT_SAVE
-    REQUIRE(omega_util_paths_equivalent(MAKE_PATH("session_save.1-2.dat"), saved_filename));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("session_save.expected.2.dat"), MAKE_PATH("session_save.1-2.dat")));
-    omega_util_remove_file(MAKE_PATH("session_save.1-3.dat"));
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("session_save.1.dat"), omega_io_flags_t::IO_FLG_NONE,
-        saved_filename));
-    REQUIRE(8 == session_events_count);// SESSION_EVT_SAVE
-    REQUIRE(omega_util_paths_equivalent(MAKE_PATH("session_save.1-3.dat"), saved_filename));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("session_save.expected.2.dat"), MAKE_PATH("session_save.1-3.dat")));
-    omega_util_remove_file(MAKE_PATH("session_save_seg.2.dat"));
-    REQUIRE(0 == omega_edit_save_segment(session_ptr, MAKE_PATH("session_save_seg.2.dat"),
-        omega_io_flags_t::IO_FLG_NONE, saved_filename, 0, 4));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("session_save_seg.expected.2.dat"),
-        MAKE_PATH("session_save_seg.2.dat")));
-    omega_util_remove_file(MAKE_PATH("session_save_seg.2.dat"));
-    omega_util_remove_file(MAKE_PATH("session_save_seg.3.dat"));
-    REQUIRE(0 == omega_edit_save_segment(session_ptr, MAKE_PATH("session_save_seg.3.dat"),
-        omega_io_flags_t::IO_FLG_NONE, saved_filename, 2, 6));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("session_save_seg.expected.3.dat"),
-        MAKE_PATH("session_save_seg.3.dat")));
-    omega_util_remove_file(MAKE_PATH("session_save_seg.3.dat"));
-    omega_edit_destroy_session(session_ptr);
-
-    // Overwrite and force overwrite tests
-    session_ptr = omega_edit_create_session(MAKE_PATH("session_save.1-3.dat"), session_save_test_session_cbk,
-                                            &session_events_count, ALL_EVENTS, nullptr);
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("session_save.1-3.dat"), omega_io_flags_t::IO_FLG_OVERWRITE,
-        saved_filename));
-    REQUIRE(omega_util_paths_equivalent(MAKE_PATH("session_save.1-3.dat"), saved_filename));
-    // overwrite twice to make sure this doesn't cause any problems
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("session_save.1-3.dat"), omega_io_flags_t::IO_FLG_OVERWRITE,
-        saved_filename));
-    REQUIRE(omega_util_paths_equivalent(MAKE_PATH("session_save.1-3.dat"), saved_filename));
-    // simulate a file being modified by another program
-    REQUIRE(0 == omega_util_touch(MAKE_PATH("session_save.1-3.dat"), 0));
-    // overwrite should fail because the original file has been modified elsewhere
-    const auto save_rc = omega_edit_save(session_ptr, MAKE_PATH("session_save.1-3.dat"),
-                                         omega_io_flags_t::IO_FLG_OVERWRITE, saved_filename);
-#ifdef OMEGA_BUILD_WINDOWS// Windows doesn't always support this
-    REQUIRE((ORIGINAL_MODIFIED == save_rc || 0 == save_rc));
-    REQUIRE((saved_filename[0] == '\0' || 0 == save_rc));
-#else
-    REQUIRE(ORIGINAL_MODIFIED == save_rc);
-    REQUIRE(saved_filename[0] == '\0');
-#endif
-    // force overwrite should succeed
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("session_save.1-3.dat"),
-        omega_io_flags_t::IO_FLG_FORCE_OVERWRITE, saved_filename));
-    REQUIRE(omega_util_paths_equivalent(MAKE_PATH("session_save.1-3.dat"), saved_filename));
-    omega_edit_destroy_session(session_ptr);
-    session_ptr = omega_edit_create_session(nullptr, nullptr, nullptr, NO_EVENTS, nullptr);
-    REQUIRE(0 == omega_edit_save(session_ptr, MAKE_PATH("session_save.empty.dat"), omega_io_flags_t::IO_FLG_OVERWRITE,
-        saved_filename));
-    REQUIRE(omega_util_paths_equivalent(MAKE_PATH("session_save.empty.dat"), saved_filename));
-    REQUIRE(0 == omega_util_compare_files(MAKE_PATH("empty_file.dat"), MAKE_PATH("session_save.empty.dat")));
-    omega_edit_destroy_session(session_ptr);
-}
-
-TEST_CASE("Transactions", "[TransactionTests]") {
-    int session_events_count = 0;
-    int viewport_events_count = 0;
-    auto session_ptr = omega_edit_create_session(nullptr, session_save_test_session_cbk, &session_events_count,
-                                                 ALL_EVENTS, nullptr);
-    REQUIRE(1 == session_events_count);// SESSION_EVT_CREATE
-    auto viewport_ptr = omega_edit_create_viewport(session_ptr, 0, 100, 0, session_save_test_viewport_cbk,
-                                                   &viewport_events_count, ALL_EVENTS);
-    REQUIRE(2 == session_events_count);// SESSION_EVT_CREATE_VIEWPORT
-    REQUIRE(1 == viewport_events_count);// VIEWPORT_EVT_CREATE
-    REQUIRE(0 == omega_session_get_transaction_state(session_ptr));
-    REQUIRE(0 == omega_viewport_get_following_byte_count(viewport_ptr));
-    auto change_id = omega_edit_insert_string(session_ptr, 0, "0123456789");
-    auto change_ptr = omega_session_get_change(session_ptr, change_id);
-    auto transaction_bit = omega_change_get_transaction_bit(change_ptr);
-    REQUIRE(0 == transaction_bit);
-    REQUIRE(0 == omega_session_get_transaction_state(session_ptr));
-    REQUIRE(3 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(2 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    REQUIRE(1 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(1 == omega_session_get_num_change_transactions(session_ptr));
-    REQUIRE(0 == omega_session_begin_transaction(session_ptr));
-    REQUIRE(1 == omega_session_get_transaction_state(session_ptr));
-    change_id = omega_edit_insert_string(session_ptr, 0, "abcdefghijklmnopqrstuvwxyz");
-    REQUIRE(2 == omega_session_get_transaction_state(session_ptr));
-    REQUIRE(4 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(3 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    change_ptr = omega_session_get_change(session_ptr, change_id);
-    REQUIRE(transaction_bit != omega_change_get_transaction_bit(change_ptr));
-    transaction_bit = omega_change_get_transaction_bit(change_ptr);
-    omega_edit_insert_string(session_ptr, 0, "ABCDEFGHIJKLMNOPQRSTUVWXYZ");
-    REQUIRE(2 == omega_session_get_transaction_state(session_ptr));
-    REQUIRE(5 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(4 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    REQUIRE(transaction_bit == omega_change_get_transaction_bit(change_ptr));
-    REQUIRE(0 == omega_session_end_transaction(session_ptr));
-    REQUIRE(3 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(2 == omega_session_get_num_change_transactions(session_ptr));
-    REQUIRE(5 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(4 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    omega_edit_insert_string(session_ptr, 0, "0123456789");
-    REQUIRE(6 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(5 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    omega_session_begin_transaction(session_ptr);
-    omega_edit_insert_string(session_ptr, 0, "abcdefghijklmnopqrstuvwxyz");
-    REQUIRE(7 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(6 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    omega_edit_overwrite_string(session_ptr, 0, "ABCDEFGHIJKLMNOPQRSTUVWXYZ");
-    REQUIRE(8 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(7 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    omega_edit_delete(session_ptr, 0, 10);
-    REQUIRE(9 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(8 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    REQUIRE(0 == omega_session_end_transaction(session_ptr));
-    REQUIRE(9 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(8 == viewport_events_count);// VIEWPORT_EVT_EDIT
-    REQUIRE(7 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(4 == omega_session_get_num_change_transactions(session_ptr));
-    REQUIRE(0 == omega_session_get_num_undone_changes(session_ptr));
-    REQUIRE(-5 == omega_edit_undo_last_change(session_ptr));
-    REQUIRE(4 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(3 == omega_session_get_num_undone_changes(session_ptr));
-    REQUIRE(3 == omega_session_get_num_change_transactions(session_ptr));
-    REQUIRE(1 == omega_session_get_num_undone_change_transactions(session_ptr));
-    REQUIRE(12 == session_events_count);// SESSION_EVT_EDIT
-    REQUIRE(7 == omega_edit_redo_last_undo(session_ptr));
-    REQUIRE(7 == omega_session_get_num_changes(session_ptr));
-    REQUIRE(0 == omega_session_get_num_undone_changes(session_ptr));
-    REQUIRE(4 == omega_session_get_num_change_transactions(session_ptr));
-
-    // Negative testing
-    REQUIRE(0 == omega_session_get_transaction_state(session_ptr));
-    REQUIRE(0 != omega_session_end_transaction(session_ptr));
-    REQUIRE(0 == omega_session_begin_transaction(session_ptr));
-    REQUIRE(1 == omega_session_get_transaction_state(session_ptr));
-    REQUIRE(0 != omega_session_begin_transaction(session_ptr));
-    REQUIRE(0 == omega_session_end_transaction(session_ptr));
-    REQUIRE(0 == omega_session_get_transaction_state(session_ptr));
-}
-
-TEST_CASE("Emoji Filename Handling", "[FilesystemTests]") {
-    const char* emoji_filenames[] = {
-        "test_😀.dat",
-        "test_👍.dat",
-        "test_🔥.dat",
-        "test 💩.dat", // Space in filename as well
-        "test_🚀.dat",
-        "test_👨‍👩‍👧‍👦.dat"  // Family emoji with zero-width joiners
-    };
-    
-    char buffer[FILENAME_MAX];
-    
-    // Test filesystem operations with emoji filenames
-    for (const auto& emoji_filename : emoji_filenames) {
-        const char dir_sep = omega_util_directory_separator();
-        std::string base_path = std::string(DATA_DIR.string().c_str()) + dir_sep;
-        std::string full_path = base_path + emoji_filename;
-        
-        // Create a file with emoji in filename
-        std::ofstream file(full_path);
-        file << "Test content" << std::endl;
-        file.close();
-        
-        // Test file_exists
-        REQUIRE(omega_util_file_exists(full_path.c_str()));
-        
-        // Test available_filename - since the file exists, we expect a path with -1 suffix
-        char* available_name = omega_util_available_filename(full_path.c_str(), buffer);
-        REQUIRE(available_name != nullptr);
-        // The file exists so available_name should append -1 to the filename
-        std::string expected_path = full_path;
-        size_t dot_pos = expected_path.rfind('.');
-        expected_path.insert(dot_pos, "-1");
-        REQUIRE(std::string(available_name) == expected_path);
-        
-        // Test basename, dirname and extension
-        char* basename_result = omega_util_basename(full_path.c_str(), nullptr, 0);
-        REQUIRE(basename_result != nullptr);
-        REQUIRE(std::string(basename_result) == emoji_filename);
-        
-        char* dirname_result = omega_util_dirname(full_path.c_str(), nullptr);
-        REQUIRE(dirname_result != nullptr);
-        REQUIRE(std::string(dirname_result) == base_path.substr(0, base_path.length() - 1));
-        
-        char* ext_result = omega_util_file_extension(full_path.c_str(), nullptr);
-        REQUIRE(ext_result != nullptr);
-        REQUIRE(std::string(ext_result) == ".dat");
-        
-        // Create a second file for copy operations
-        std::string copy_path = base_path + "copy_" + emoji_filename;
-        
-        // Test file_copy
-        REQUIRE(0 == omega_util_file_copy(full_path.c_str(), copy_path.c_str(), 0));
-        REQUIRE(omega_util_file_exists(copy_path.c_str()));
-        REQUIRE(0 == omega_util_compare_files(full_path.c_str(), copy_path.c_str()));
-        
-        // Test remove_file
-        REQUIRE(0 == omega_util_remove_file(full_path.c_str()));
-        REQUIRE(!omega_util_file_exists(full_path.c_str()));
-        
-        REQUIRE(0 == omega_util_remove_file(copy_path.c_str()));
-        REQUIRE(!omega_util_file_exists(copy_path.c_str()));
-    }
-}
